collector package refactoring
