+++ server/collector/src/main/java/org/bithon/server/collector/source/brpc/BrpcEventCollector.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.collector.brpc;
+package org.bithon.server.collector.source.brpc;
 
 import lombok.extern.slf4j.Slf4j;
 import org.bithon.agent.rpc.brpc.BrpcMessageHeader;
