+++ server/collector/src/main/java/org/bithon/server/collector/source/kafka/KafkaCollectorStarter.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.collector.kafka;
+package org.bithon.server.collector.source.kafka;
 
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.common.serialization.StringDeserializer;
